+++ MIPT/problem_HWH/hashtable.h
@@ -4,7 +4,7 @@
 typedef struct hashtable_t Hashtable;
 
 void *malloc_wrap(size_t size);
-Hashtable *init(Hashtable *table, unsigned long size,
+Hashtable *init(Hashtable *table, unsigned size,
   unsigned long (*hash)(const char *str));
 void add2table(Hashtable *table, const char *token);
 void count_words(Hashtable *table, const char *token);
